move config to envs
