+++ src/components/Analytics/Google.tsx
@@ -1,6 +1,6 @@
 import { GoogleAnalytics as GA } from '@next/third-parties/google';
 
-import { analyticsEnv } from '@/config/analytics';
+import { analyticsEnv } from '@/envs/analytics';
 
 const GoogleAnalytics = () => <GA gaId={analyticsEnv.GOOGLE_ANALYTICS_MEASUREMENT_ID!} />;
