+++ src/components/Analytics/Vercel.tsx
@@ -1,7 +1,7 @@
 import { Analytics } from '@vercel/analytics/react';
 import { memo } from 'react';
 
-import { analyticsEnv } from '@/config/analytics';
+import { analyticsEnv } from '@/envs/analytics';
 
 const VercelAnalytics = memo(() => <Analytics debug={analyticsEnv.DEBUG_VERCEL_ANALYTICS} />);
