+++ src/libs/traces/index.ts
@@ -1,8 +1,8 @@
 import { Langfuse } from 'langfuse';
 import { CreateLangfuseTraceBody } from 'langfuse-core';
 
-import { getLangfuseConfig } from '@/config/langfuse';
 import { CURRENT_VERSION } from '@/const/version';
+import { getLangfuseConfig } from '@/envs/langfuse';
 import { TraceEventClient } from '@/libs/traces/event';
 
 /**
