+++ src/server/services/search/impls/searxng/index.ts
@@ -1,6 +1,6 @@
 import { TRPCError } from '@trpc/server';
 
-import { toolsEnv } from '@/config/tools';
+import { toolsEnv } from '@/envs/tools';
 import { SearXNGClient } from '@/server/services/search/impls/searxng/client';
 import { SEARCH_SEARXNG_NOT_CONFIG, UniformSearchResponse } from '@/types/tool/search';
